adds two hangman input parsing methods to Parser and integrates existing Parser methods into Hangman
